requirements.txt
pyyaml
tqdm
numpy
scikit-image >= 0.19.0
matplotlib
extorch >= 1.0.4
tensorboard
torch >= 1.4.0
torchvision >= 0.4.0
